update app.test.tsx and app.tsx
Changed behavior of dropzone to not include file previews, because
interacting with them wasn't purposeful. On adding a file to the
dropzone, it was automatically uploaded, vs requiring a button press, so
we opted to remove the preview (and set fileLimit = -1) so that it feels
more natural to upload a file. Bonus: No error from uploading more than
3 files. Added a little styling around the dropzone, marginY=32px.

closes #26